+++ src/Raven.CodeAnalysis/Syntax/InternalSyntax/Parser/Parsers/EnumDeclarationParser.cs
@@ -121,6 +121,7 @@ internal class EnumDeclarationParser : SyntaxParser
                      SyntaxKind.PrivateKeyword or
                      SyntaxKind.InternalKeyword or
                      SyntaxKind.ProtectedKeyword or
+                     SyntaxKind.FilescopeKeyword or
                      SyntaxKind.StaticKeyword or
                      SyntaxKind.AbstractKeyword or
                      SyntaxKind.FinalKeyword or